Create README.md
I am trying to be the best person I can be. I hope that all the different opportunities i have been working on that I will be able to help people become the best person they can be. life is way to short not to following dreams and better opportunitis to live a life and not having to work 2 jobs to make ends meet.